Perform refactor
