+++ lib/boombox/storage_endpoints/aac.ex
@@ -1,0 +1,33 @@
+defmodule Boombox.StorageEndpoints.AAC do
+  @moduledoc false
+  import Membrane.ChildrenSpec
+  alias Boombox.StorageEndpoints
+  alias Boombox.Pipeline.Ready
+  alias Membrane.AAC
+
+  @spec create_input(String.t(), transport: :file | :http) :: Ready.t()
+  def create_input(location, opts) do
+    spec =
+      StorageEndpoints.get_source(location, opts[:transport])
+      |> child(:aac_input_parser, Membrane.AAC.Parser)
+
+    %Ready{track_builders: [{:audio, spec}]}
+  end
+
+  @spec link_output(
+          String.t(),
+          Boombox.Pipeline.track_builders(),
+          Membrane.ChildrenSpec.t()
+        ) :: Ready.t()
+  def link_output(location, track_builders, _spec_builder) do
+    spec =
+      StorageEndpoints.get_track(track_builders, :audio)
+      |> child(:aac_audio_transcoder, %Membrane.Transcoder{
+        output_stream_format: %AAC{}
+      })
+      |> child(:aac_output_parser, Membrane.AAC.Parser)
+      |> child(:file_sink, %Membrane.File.Sink{location: location})
+
+    %Ready{actions: [spec: spec]}
+  end
+end
